refactor: move Psych-DS utility functions from CLI to core metadata package
isValidPsychDSDataFilename, toPsychDSValue, deriveArrayFilename,
objectsToCSV, and disambiguateArrayFilename are pure functions with no
platform dependencies. Moving them into @jspsych/metadata makes them
available to any downstream consumer (frontend, tests, etc.) without
needing to reach into CLI internals or duplicate the logic.

CLI now imports these from @jspsych/metadata; Node-specific helpers
(expandHomeDir, fileStem, disambiguateFilename) remain in cli/src/utils.ts.
